Refactor GameObject to module
