+++ pyproject.toml
@@ -14,9 +14,15 @@ dev = [
 ]
 
 [tool.ruff.lint]
-select = ["I"]
+select = [
+    "ANN201",  # Missing return type annotation for public function
+    "ANN204",  # Missing return type annotation for special method
+    "I",
+]
 
 [tool.ruff.lint.per-file-ignores]
 "CondaRTS.py" = [
+    "ANN201",
+    "ANN204",
     "COM812",  # Trailing comma missing - conflicts with ruff format
 ]
