+++ src/constants.py
@@ -1,0 +1,3 @@
+"""Constants used by `src/` modules."""
+
+MAP_WIDTH, MAP_HEIGHT = 1600, 800
